Module oak_virtio::queue::virtq

Structs

AvailRing

The ring buffer that indicates which descriptors have been made available by the driver.

Desc

A descriptor for a byte buffer used in a virtio queue.

DescFlags

Flags about a descriptor.

RingFlags

Flags about the available and used rings.

UsedElem

An element indicating a used descriptor chain.

UsedRing

The ring buffer that indicates which available descriptors have been consumed by the device.

VirtQueue

A split virtqueue implementation.