Struct oak_virtio::queue::virtq::Desc

#[repr(C, align(16))]
pub struct Desc {
    pub addr: PhysAddr,
    pub length: u32,
    pub flags: DescFlags,
    pub next: u16,
}

A descriptor for a byte buffer used in a virtio queue.

See https://docs.oasis-open.org/virtio/virtio/v1.1/csprd01/virtio-v1.1-csprd01.html#x1-320005.

Fields
addr: PhysAddr

The guest-physical address of the buffer.

length: u32

The lengths of the buffer.

flags: DescFlags

Flags providing more info about this descriptor.

next: u16

The index of the next descriptor in the chain if this is not the end (i.e. flags contain DescFlags::VIRTQ_DESC_F_NEXT), ignored otherwise.

Not used for now, as we don’t currently support chaining.

Implementations

impl Desc

pub fn new(flags: DescFlags, addr: PhysAddr, length: u32) -> Self