Trait oak_virtio::Read

pub trait Read {
    // Required method
    fn read_exact(&mut self, data: &mut [u8]) -> Result<()>;
}

Read bytes from a source.

This trait is similar to the std::io::Read trait, except that this trait is pared down to a minimum and works in a no_std environment.

Required Methods

fn read_exact(&mut self, data: &mut [u8]) -> Result<()>

Read bytes until data has been filled.